+++ lib/graphql_rails/attributes/type_parser.rb
@@ -9,38 +9,13 @@ module GraphqlRails
     # converts string value in to GraphQL type
     class TypeParser
       require_relative './type_name_info'
+      require_relative './type_parseable'
 
-      class UnknownTypeError < ArgumentError; end
       class NotSupportedFeature < GraphqlRails::Error; end
 
-      TYPE_MAPPING = {
-        'id' => GraphQL::ID_TYPE,
+      include TypeParseable
 
-        'int' => GraphQL::INT_TYPE,
-        'integer' => GraphQL::INT_TYPE,
-
-        'string' => GraphQL::STRING_TYPE,
-        'str' => GraphQL::STRING_TYPE,
-        'text' => GraphQL::STRING_TYPE,
-        'time' => GraphQL::STRING_TYPE,
-        'date' => GraphQL::STRING_TYPE,
-
-        'bool' => GraphQL::BOOLEAN_TYPE,
-        'boolean' => GraphQL::BOOLEAN_TYPE,
-
-        'float' => GraphQL::FLOAT_TYPE,
-        'double' => GraphQL::FLOAT_TYPE,
-        'decimal' => GraphQL::FLOAT_TYPE
-      }.freeze
-
-      RAW_GRAPHQL_TYPES = [
-        GraphQL::Schema::List,
-        GraphQL::BaseType,
-        GraphQL::ObjectType,
-        GraphQL::InputObjectType
-      ].freeze
-
-      delegate :list?, :required_inner_type?, :required_list?, :nullable_inner_name, :required?, to: :type_name_info
+      delegate :list?, :required_inner_type?, :required_list?, :required?, to: :type_name_info
 
       def initialize(unparsed_type, paginated: false)
         @unparsed_type = unparsed_type
@@ -71,15 +46,6 @@ module GraphqlRails
         end
       end
 
-      def graphql_model
-        type_class = nullable_inner_name.safe_constantize
-
-        return if type_class.nil?
-        return unless type_class < GraphqlRails::Model
-
-        type_class
-      end
-
       protected
 
       def paginated_type_arg
@@ -103,22 +69,7 @@ module GraphqlRails
       end
 
       def raw_unwrapped_type
-        @raw_unwrapped_type ||= begin
-          type = parsed_type
-          type = type.of_type while wrapped_type?(type)
-          type
-        end
-      end
-
-      def wrapped_type?(type)
-        type.is_a?(GraphQL::ListType) || type.is_a?(GraphQL::NonNullType) || type.is_a?(GraphQL::Schema::List)
-      end
-
-      def dynamicly_defined_type
-        type_class = graphql_model
-        return unless type_class
-
-        type_class.graphql.graphql_type
+        @raw_unwrapped_type ||= unwrap_type(parsed_type)
       end
 
       private
@@ -143,14 +94,6 @@ module GraphqlRails
         end
       end
 
-      def raw_graphql_type?
-        return true if RAW_GRAPHQL_TYPES.detect { |raw_type| unparsed_type.is_a?(raw_type) }
-
-        defined?(GraphQL::Schema::Member) &&
-          unparsed_type.is_a?(Class) &&
-          unparsed_type < GraphQL::Schema::Member
-      end
-
       def type_name_info
         @type_name_info ||= begin
           type_name = \
@@ -164,13 +107,14 @@ module GraphqlRails
       end
 
       def type_by_name
-        TYPE_MAPPING.fetch(nullable_inner_name.downcase) do
-          dynamicly_defined_type || raise(
-            UnknownTypeError,
-            "Type #{unparsed_type.inspect} is not supported. Supported scalar types are: #{TYPE_MAPPING.keys}." \
-            ' All the classes that includes `GraphqlRails::Model` are also supported as types.'
-          )
-        end
+        unwrapped_scalar_type || unwrapped_model_type || raise_not_supported_type_error
+      end
+
+      def unwrapped_model_type
+        type_class = graphql_model
+        return unless type_class
+
+        type_class.graphql.graphql_type
       end
     end
   end